% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{Background_sampling}
\alias{Background_sampling}
\title{Background sampling}
\usage{
Background_sampling(
  counts_norm_mat,
  cell_clusterid_mat,
  ncells.sample = 100,
  clusterids_vec = "cell.type",
  min.cells.gexp = 10
)
}
\arguments{
\item{counts_norm_mat}{normalised umi counts matrix-- genes X UMIs}

\item{cell_clusterid_mat}{matrix with columns --"cell.name(UMI)"--same as colnames of counts matrix,"cluster.id","Tissue"}

\item{ncells.sample}{number of cells to sample}

\item{clusterids_vec}{vector with elements corresponding column names in cell_clusterid_mat with the correct heirarchiel order of clusterids (eg. Tissue, cell type, cell subtype)}

\item{min.cells.gexp}{minimum of cells genes need to be expressed in}
}
\value{
sampled background
}
\description{
Background sampling
}
