% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{cells_exp}
\alias{cells_exp}
\title{Filter cells with non zero gene expression}
\usage{
cells_exp(data_matrix, min.cells.gexp)
}
\arguments{
\item{data_matrix}{matrix genes x UMIs}

\item{min.cells.gexp}{minimum of cells genes need to be expressed in}
}
\value{
list of genes with names of cells (UMIs) that have non zero gene exp values
}
\description{
Filter cells with non zero gene expression
}
