% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{cleanData}
\alias{cleanData}
\title{Data cleaning: removing genes not expressed, removing outliers & genes for which we cannot properly approximate distribution}
\usage{
cleanData(data_counts, dir.name, file.id)
}
\arguments{
\item{data_counts}{single cell matrix}

\item{dir.name}{directory where to save data}

\item{file.id}{prefix to use when saving data}
}
\value{
cleaned data
}
\description{
Data cleaning: removing genes not expressed, removing outliers & genes for which we cannot properly approximate distribution
}
\author{
Céline Barlier
}
