% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{compute_pval}
\alias{compute_pval}
\title{Compute p-values}
\usage{
compute_pval(th_list, gexp_mat, genes, p_correct = T)
}
\arguments{
\item{th_list}{list of background thresholds}

\item{gexp_mat}{gene expression matrix (query)}

\item{genes}{genes for which p-values will be computed}

\item{p_correct}{perform multiple correction}
}
\value{
p-values (corrected if p_correct = T)
}
\description{
Compute p-values
}
