% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{Create.sig.frame}
\alias{Create.sig.frame}
\title{Function to quantify gene level at the population level and identify key identity genes}
\usage{
Create.sig.frame(discrete.mat)
}
\arguments{
\item{discrete.mat}{discretized matrix}
}
\value{
data.frame of gene levels frequency
}
\description{
Function to quantify gene level at the population level and identify key identity genes
}
