% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{createEcdfsBack}
\alias{createEcdfsBack}
\title{Generate ecdf function for each gene}
\usage{
createEcdfsBack(gexp, swap = TRUE)
}
\arguments{
\item{gexp}{gene expression vector}
}
\value{
ecdfs
}
\description{
Generate ecdf function for each gene
}
\author{
Céline Barlier
}
