% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{createThresholdDist}
\alias{createThresholdDist}
\title{Generate threshold distributions}
\usage{
createThresholdDist(gexp, swap = TRUE)
}
\arguments{
\item{gexp}{gene expression vector}

\item{swap}{parameter to ensure consistency of lower & upper-bounds}
}
\value{
thresholds
}
\description{
Generate threshold distributions
}
