% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{DataNorm}
\alias{DataNorm}
\title{Data Normalisation}
\usage{
DataNorm(
  data,
  log_transform = c(TRUE, FALSE),
  data.type,
  bg.model.pars = NULL,
  file.id = NULL
)
}
\arguments{
\item{data}{single cell data to normalize}

\item{log_transform}{log transformation}

\item{data.type}{type of data (Ref or query)}

\item{bg.model.pars}{background factors retrieved from scTransform (used when normalizing query)}

\item{file.id}{presfix to use when saving files}
}
\value{
normalized data
}
\description{
Data Normalisation
}
