% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{discrete.expr.levels}
\alias{discrete.expr.levels}
\title{Function to determine the gene expression level of key identity genes}
\usage{
discrete.expr.levels(tbl, zs)
}
\arguments{
\item{tbl}{sig.frame data.frame obtained from Create.sig.frame()}

\item{zs}{z-score computed in Create.sig.frame()}
}
\value{
key identity genes data.frame (with gene quantification level)
}
\description{
Function to determine the gene expression level of key identity genes
}
\author{
Céline Barlier
}
