% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{Discretise_data}
\alias{Discretise_data}
\title{Discretize gene expression in three levels of expression: 0 (not expressed), 0.5 (medium level), 1 (high level of expression)}
\usage{
Discretise_data(
  query_ScaledData,
  Ref_data,
  thsback = NA,
  th = 0.05,
  th_i = 0.05,
  file.id = "myquery",
  return.data = T,
  saveData = T
)
}
\arguments{
\item{query_ScaledData}{prepared query data}

\item{Ref_data}{background reference}

\item{thsback}{background thresholds}

\item{th}{threshold cutoffs}

\item{th_i}{threshold cutoffs}

\item{file.id}{prefix to use when saving files}

\item{return.data}{return results}

\item{save}{data in rds object}
}
\value{
discretized gene expression matrix
}
\description{
Discretize gene expression in three levels of expression: 0 (not expressed), 0.5 (medium level), 1 (high level of expression)
}
