% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{.optHighFun}
\alias{.optHighFun}
\title{Optimization function to compute upper-bound thresholds}
\usage{
.optHighFun(x, empcdf, maxVal)
}
\description{
Optimization function to compute upper-bound thresholds
}
