% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{.optLowFun}
\alias{.optLowFun}
\title{Optimization function to compute lower-bound thresholds}
\usage{
.optLowFun(x, empcdf, maxVal)
}
\description{
Optimization function to compute lower-bound thresholds
}
