% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{get.identity.genes}
\alias{get.identity.genes}
\title{Function to get identity genes: high level of expression (from discrete.expr.levels)}
\usage{
get.identity.genes(tbl, bimodg)
}
\arguments{
\item{tbl}{sig.frame data.frame obtained from Create.sig.frame()}

\item{bimodg}{genes with a bimodal distribution in the background (used to identify "unique" mediumly expressed genes)}
}
\value{
data.frame of identity genes and their level (high/medium)
}
\description{
Function to get identity genes: high level of expression (from discrete.expr.levels)
}
\author{
Céline Barlier
}
