% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{GetBackground}
\alias{GetBackground}
\title{Background sampling & threshold computation}
\usage{
GetBackground(
  data_norm,
  dir_name,
  file.id,
  metadata,
  sample_n,
  parallelize,
  ncores,
  num.boot.sample,
  fixseed
)
}
\arguments{
\item{data_norm}{normalized single cell matrix}

\item{dir_name}{directory name where to save the data}

\item{file.id}{prefix to use when saving data}

\item{metadata}{metadata contains cells.id & cell.type}

\item{sample_n}{number of cells to use by group for sampling}

\item{parallelize}{run the code in parallel}

\item{ncores}{number of cores to use if parallelize = T}

\item{num.boot.sample}{number of bootstrap samples to run}

\item{fixseed}{seed used to generate ecdfBack}
}
\value{
background thresholds
}
\description{
Background sampling & threshold computation
}
\author{
Céline Barlier
}
