% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{GetBimodalGenes}
\alias{GetBimodalGenes}
\title{Identify genes with a biomdal distribution in the background}
\usage{
GetBimodalGenes(
  runi,
  data_norm,
  dir_name,
  file.id,
  metadata,
  sample_n,
  ncores,
  parallelize = TRUE
)
}
\arguments{
\item{data_norm}{normalized single cell matrix}

\item{dir_name}{directory name where to save the data}

\item{file.id}{prefix to use when saving data}

\item{metadata}{metadata contains cells.id & cell.type}

\item{sample_n}{number of cells to use by group for sampling}

\item{ncores}{number of cores to use if parallelize = Ts}

\item{parallelize}{run the code in parallel}
}
\value{
background thresholds
}
\description{
Identify genes with a biomdal distribution in the background
}
\author{
Céline Barlier
}
