% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{GetEcdfs}
\alias{GetEcdfs}
\title{Get ecdfs of the background}
\usage{
GetEcdfs(processed.back, file.id)
}
\arguments{
\item{processed.back}{sampled & sclaed single cell matrix}

\item{file.id}{prefix for the file to save}
}
\value{
sampled background
}
\description{
Get ecdfs of the background
}
\author{
Modified by Céline Barlier
}
