% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{PrepQueryData}
\alias{PrepQueryData}
\title{Prepare query data: clean, normalize & scale}
\usage{
PrepQueryData(
  query_dat,
  Ref_data,
  dir,
  sfback = NA,
  file.id = "query",
  ncores,
  bg.model.pars,
  normalise = T,
  scale_data = T,
  saveData = T,
  return.data = T
)
}
\arguments{
\item{query_dat}{query data}

\item{Ref_data}{background reference}

\item{dir}{directory used to save the results}

\item{sfback}{scaling factors}

\item{file.id}{prefix to use when saving files}

\item{ncores}{number of cores to use to run the code}

\item{bg.model.pars}{background factors retrieved from scTransform (used when normalizing query)}

\item{normalise}{normalize data}

\item{scale_data}{scale data}

\item{return.data}{return results}

\item{save}{data in rds object}
}
\value{
prepared query data
}
\description{
Prepare query data: clean, normalize & scale
}
