% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{RunBackComputations}
\alias{RunBackComputations}
\title{Function called by RunBootstrap() to sample, scale & compute the background thresholds}
\usage{
RunBackComputations(
  norm.data,
  cell_clusterid_mat,
  ncells.sample,
  clusterids_vec,
  file.id
)
}
\arguments{
\item{norm.data}{normalized single cell matrix}

\item{cell_clusterid_mat}{metadata}

\item{ncells.sample}{number of cells to use by group for the sampling}

\item{clusterids_vec}{metadata column name with the cell type info}

\item{file.id}{prefix for the file to save}
}
\value{
sampled background
}
\description{
Function called by RunBootstrap() to sample, scale & compute the background thresholds
}
\author{
Modified by Céline Barlier
}
