% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{RunBootstrap}
\alias{RunBootstrap}
\title{Run Bootstrapping for the sampling, scalling & thresholds computation of the background}
\usage{
RunBootstrap(
  data_norm,
  file.id,
  clusterids_vec,
  metadata,
  sample_n,
  parallelize,
  ncores,
  num.boot.sample,
  fixseed
)
}
\arguments{
\item{data_norm}{normalized single cell matrix}

\item{file.id}{prefix to use when saving data}

\item{metadata}{metadata contains cells.id & cell.type}

\item{sample_n}{number of cells to use by group for sampling}

\item{parallelize}{run the code in parallel}

\item{ncores}{number of cores to use if parallelize = T}

\item{num.boot.sample}{number of bootstrap samples to run}

\item{fixseed}{seed used to generate ecdfBack}

\item{saveData}{save data in the folder}
}
\value{
bootstrapped thresholds & ecdfs functions
}
\description{
Run Bootstrapping for the sampling, scalling & thresholds computation of the background
}
\author{
Céline Barlier
}
