% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{sample.data}
\alias{sample.data}
\title{Sample cells}
\usage{
sample.data(cells_dat, cellid, ncells.sample, nlevels)
}
\arguments{
\item{cells_dat}{vector of cell names/UMI}

\item{cellid}{named vector (names=cell name/UMI, elements=clusterid-combined id)}

\item{ncells.sample}{number of cells to sample}

\item{nlevel}{number of cluster id levels (e.g Tissue , cell type, cell subtype)}
}
\value{
sampled cells
}
\description{
Sample cells
}
