% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{ScaleBackground}
\alias{ScaleBackground}
\title{Scaling of the background}
\usage{
ScaleBackground(sampled.back, file.id)
}
\arguments{
\item{sampled.back}{sampled single cell matrix}

\item{file.id}{prefix for the file to save}
}
\value{
sampled background
}
\description{
Scaling of the background
}
\author{
Modified by Céline Barlier
}
