% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{ScaledData}
\alias{ScaledData}
\title{Scaling of data}
\usage{
ScaledData(data)
}
\arguments{
\item{data}{single cell matrix to scale with maximum gene expression}
}
\value{
scaled matrix
}
\description{
Scaling of data
}
