% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{scaleMaxGexp}
\alias{scaleMaxGexp}
\title{Called by ScaledData()}
\usage{
scaleMaxGexp(data_mat, scalingfactors, genes)
}
\arguments{
\item{data_mat}{single cell matrix}

\item{scalingfactors}{scaling factors}

\item{genes}{gene names}
}
\value{
scaled expression
}
\description{
Called by ScaledData()
}
