% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{scrb_norm}
\alias{scrb_norm}
\title{Data normalization UMI}
\usage{
scrb_norm(
  data_counts,
  log_transform = T,
  bg.model.pars = NULL,
  data.type = c("Ref", "query"),
  file.id = "mydata",
  saveData = T,
  return = T
)
}
\arguments{
\item{data_counts}{single cell matrix}

\item{log_transform}{log transformation}

\item{bg.model.pars}{background factors retrieved from scTransform (used when normalizing query)}

\item{data.type}{type of data: Ref (background) or query}

\item{file.id}{prefix to use when saving files}

\item{saveData}{save normalized data in rds object}

\item{return}{return normalized data from the function}
}
\value{
normalized data
}
\description{
Data normalization UMI
}
