% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{swap_th}
\alias{swap_th}
\title{Verify consistency of computed thresholds
Check that lower gexp value are lower thresholds & higher gexp value are upper thresholds}
\usage{
swap_th(th_dist)
}
\arguments{
\item{th_dist}{thresholds}
}
\value{
thresolds
}
\description{
Verify consistency of computed thresholds
Check that lower gexp value are lower thresholds & higher gexp value are upper thresholds
}
